refactor(ui): split AccessRequestBanner into focused files, add sliding pill and shared avatar animation
Major improvements:
- Split monolithic file into types.ts, decision-pill.tsx, summary-text.tsx,
  access-request-banner.tsx, and index.ts barrel export
- Summary text now highlights @handles and "N others" in foreground color
  while keeping filler text muted
- Avatar stack now left-to-right (leftmost on top) via flex-row-reverse
- DecisionPill redesigned with sliding highlight indicator that animates
  between Approve/Reject positions using layoutId
- Collapsed avatars conditionally unmount when expanded so layoutId shared
  animation can bounce avatars from header into their expanded row positions
- Expanded text crossfades to "Access Requests" title when opened
- Summary text allows 2-line wrap (line-clamp-2) for narrow widths
- Bumped EXPAND_SPRING damping to 22 to fix weird fold animation cutoff
- cursor-pointer on all interactive elements
