+++ frontend/components/access-request-banner/index.ts
@@ -1,0 +1,5 @@
+export { AccessRequestBanner } from "./access-request-banner";
+export type {
+	AccessRequest,
+	AccessRequestBannerProps,
+} from "./types";
